+++ CHANGELOG.md
@@ -40,7 +40,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - VSTi queries the host sample rate more robustly. Cubase previously reported
   the sample rate as 0 Hz, leading to persistent error message about the sample
   rate not being 44100 Hz. ([#222][i222])
-- Occasional NaNs in the Trisaw oscillator when the color was 0 in the Go VM.
+- Occasional NaNs in the Trisaw oscillator when color was = 0 or color = 128
 - The tracker thought that "sync" unit pops the value from stack, even if the VM
   did not, resulting it claiming errors in patches that worked once compiled.